requirements.txt
# Core Framework
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Generative AI - Google Gemini
google-generativeai==0.3.0

# Utilities
markdown==3.5.1